pyproject.toml
[project]
name = "devr-ai"
version = "0.1.0"
description = "Devr.AI is an AI-powered Developer Relations (DevRel) assistant designed to seamlessly integrate with open-source communities across platforms like Discord, Slack, GitHub, and Discourse."
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "supabase (>=2.13.0,<3.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "py-cord (>=2.6.1,<3.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "slack-sdk (>=3.34.0,<4.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langgraph (>=0.4.7,<0.5.0)",
    "langchain-tavily (>=0.2.0,<0.3.0)",
    "tavily-python (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langsmith (>=0.3.45,<0.4.0)",
    "weaviate-client (>=4.15.0,<5.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "ddgs (>=9.0.2,<10.0.0)",
    "fastmcp>=2.11.3,<3.0.0",
    "PyGithub>=2.5.0,<3.0.0",
    "discord-py (>=2.5.2,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "devr" }]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
flake8 = "^7.1.2"
autopep8 = "^2.3.2"
autoflake = "^2.3.1"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
